Extract some commands to admin cog
