+++ bot/main.py
@@ -52,6 +52,7 @@ class RedstoneSquid(Bot):
         await self.load_extension("jishaku")
         await self.load_extension("bot.verify")
         await self.load_extension("bot.delete_log")
+        await self.load_extension("bot.admin")
 
 
 async def main():
